Break executor source and some other changes
