+++ crates/core/src/wasm_runtime.rs
@@ -4,20 +4,20 @@ mod delegate;
 mod delegate_store;
 mod error;
 mod native_api;
+mod runtime;
 mod secrets_store;
 mod state_store;
 mod store;
 #[cfg(test)]
 mod tests;
-mod wasm_runtime;
 
 pub(crate) use contract::ContractRuntimeInterface;
 pub use contract_store::ContractStore;
 pub(crate) use delegate::DelegateRuntimeInterface;
 pub use delegate_store::DelegateStore;
 pub(crate) use error::{ContractError, RuntimeInnerError, RuntimeResult};
+pub use runtime::{ContractExecError, Runtime};
 pub(crate) use secrets_store::SecretStoreError;
 pub use secrets_store::SecretsStore;
 pub use state_store::StateStore;
 pub(crate) use state_store::{StateStorage, StateStoreError};
-pub use wasm_runtime::{ContractExecError, Runtime};
