+++ crates/core/src/wasm_runtime/contract.rs
