+++ crates/core/src/wasm_runtime/contract_store.rs
