+++ crates/core/src/wasm_runtime/delegate.rs
